morphing_pipeline/pyproject.toml
[project]
name = "morphing-pipeline"
version = "0.1.0"
description = "3D mesh morphing pipeline: image/mesh → GPT morph → 3D target → transition"
requires-python = "==3.12.*"
dependencies = [
    "trimesh[easy]>=4.4.0,<5.0",
    "scipy>=1.13,<2.0",
    "numpy>=1.26,<2.0",
    "matplotlib>=3.9",
    "gradio-client>=1.3.0",
    "requests>=2.32",
    "pyyaml>=6.0.3",
    "pillow>=10.0",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
    "huggingface-hub>=0.23",
    "langgraph>=0.2,<1.0",
    "typing-extensions>=4.9",
    "openai>=1.30",
    "scikit-image>=0.26.0",
    "fast-simplification>=0.1.13",
]

[project.optional-dependencies]
# pyrender + OSMesa: used in Docker (Linux). Install with: uv sync --extra render
render = [
    "pyrender>=0.1.45",
    "PyOpenGL>=3.1.0",
]
# PyMeshLab for stage-1 mesh repair. Install with: uv sync --extra repair
repair = [
    "pymeshlab>=2023.12",
]
# PyTorch for differential rendering morphing (node 7b). Install with: uv sync --extra diffrender
# NOTE: pytorch3d must be installed separately (not on PyPI) — see instructions below:
#   CPU dev (macOS/Linux):
#     pip install fvcore iopath
#     pip install --no-build-isolation "git+https://github.com/facebookresearch/pytorch3d.git@stable"
#   Docker GPU (add to Dockerfile after pytorch/pytorch base image):
#     RUN pip install fvcore iopath
#     RUN pip install --no-build-isolation "git+https://github.com/facebookresearch/pytorch3d.git@stable"
diffrender = [
    "torch>=2.2",
    "fvcore>=0.1.5",
    "iopath>=0.1.9",
]

[tool.setuptools]
py-modules = ["run_morphing"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
